RavenDB-25281 Corax: dedupe search-term splitting into QueryBuilderHelper (#4828)
LuceneQueryBuilder.HandleSearch had three local functions (GetValues, YieldValue,
IsEscaped) that split a search value on whitespace honoring "..." phrases and
\-escapes. QueryPlanBuilder.Resolution.cs had a near-duplicate (SplitSearchTerms +
StripEscapes) — same intent, slightly less correct escape handling.

Extracted the Lucene implementation to QueryBuilderHelper.SplitSearchValue and
pointed both callers at it. Validated with SlowTests.Issues.RavenDB_15924 (4
sub-tests pass).
