+++ src/Raven.Server/Documents/Queries/LuceneQueryBuilder.cs
@@ -4,7 +4,6 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Runtime.CompilerServices;
-using System.Text;
 using Lucene.Net.Analysis;
 using Lucene.Net.Search;
 using Lucene.Net.Spatial.Queries;
@@ -956,7 +955,7 @@ namespace Raven.Server.Documents.Queries
             BooleanQuery q = null;
             var occur = Occur.SHOULD;
             Lucene.Net.Search.Query firstQuery = null;
-            foreach (var v in GetValues())
+            foreach (var v in QueryBuilderHelper.SplitSearchValue(valueAsString))
             {
                 if (LuceneQueryHelper.TryGetAnalyzedTerm(fieldName, v, GetTermType(v), analyzer, out var t, index) == false)
                     continue;
@@ -1018,104 +1017,6 @@ namespace Raven.Server.Documents.Queries
 
                 return LuceneTermType.String;
             }
-
-            /*
-             * Here we need to deal with value that comes from the user, which means that we
-             * have to be careful.
-             *
-             * The rules are that we'll split the terms on whitespace, except if they are quoted
-             * using ", however, you may escape the " using \, and \ using \\.
-             */
-            IEnumerable<string> GetValues()
-            {
-                List<int> escapePositions = null;
-
-                var quoted = false;
-                var lastWordStart = 0;
-                for (var i = 0; i < valueAsString.Length; i++)
-                {
-                    switch (valueAsString[i])
-                    {
-                        case '\\' when IsEscaped(valueAsString, i):
-                            AddEscapePosition(i);
-                            break;
-                        case '"':
-                            if (IsEscaped(valueAsString, i))
-                            {
-                                AddEscapePosition(i);
-                                continue;
-                            }
-
-                            if (lastWordStart != i)
-                            {
-                                yield return YieldValue(valueAsString, lastWordStart, i - lastWordStart, escapePositions);
-                            }
-
-                            quoted = !quoted;
-                            lastWordStart = i + 1;
-                            break;
-                        case '\t':
-                        case ' ':
-                            if (quoted)
-                                continue;
-
-                            if (lastWordStart != i)
-                            {
-                                yield return YieldValue(valueAsString, lastWordStart, i - lastWordStart, escapePositions);
-                            }
-
-                            lastWordStart = i + 1; // skipping
-                            break;
-                    }
-                }
-
-                if (valueAsString.Length - lastWordStart > 0)
-                    yield return YieldValue(valueAsString, lastWordStart, valueAsString.Length - lastWordStart, escapePositions);
-
-
-
-                void AddEscapePosition(int i)
-                {
-                    escapePositions ??= new List<int>(16);
-                    escapePositions.Add(i - 1);
-                }
-            }
-
-            string YieldValue(string input, int startIndex, int length, List<int> escapePositions)
-            {
-                if (escapePositions == null || escapePositions.Count == 0)
-                    return input.Substring(startIndex, length);
-
-                var sb = new StringBuilder(input, startIndex, length, length);
-
-                for (int i = escapePositions.Count - 1; i >= 0; i--)
-                {
-                    sb.Remove(escapePositions[i] - startIndex, 1);
-                }
-
-                escapePositions.Clear();
-
-                return sb.ToString();
-            }
-
-            bool IsEscaped(string input, int index)
-            {
-                var count = 0;
-                for (int i = index - 1; i >= 0; i--)
-                {
-                    if (input[i] == '\\')
-                    {
-                        count++;
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-
-                return (count & 1) == 1;
-            }
-
         }
 
         private static Lucene.Net.Search.Query HandleSpatial(Query query, MethodExpression expression, QueryMetadata metadata, BlittableJsonReaderObject parameters,
